some tests has been added

+++ src/main/java/equation/app/model/Answer.java
@@ -8,6 +8,7 @@ import jakarta.persistence.JoinColumn;
 import jakarta.persistence.JoinTable;
 import jakarta.persistence.ManyToMany;
 import jakarta.persistence.Table;
+import java.math.BigDecimal;
 import java.util.List;
 import lombok.Data;
 
@@ -18,7 +19,7 @@ public final class Answer {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
-    private Double value;
+    private BigDecimal value;
     @ManyToMany
     @JoinTable(name = "equations_answers",
             joinColumns = @JoinColumn(name = "answer_id"),
